move some useful common code from combined_database.cpp
